Fix the argument passing since Kitsune can correctly pass the GPU architecture all the way down to LLD now.
Resurrect the euler3d multi-file test and ensure that it checks its output
correctly

Fix the vecadd multi-source test to use the "interface" code from the vecadd
single-source test.

Enable globbing in the multi-source directory, but add the ability to exclude
certain directories. This is temporary because we have code in here that does
not currently work.

Ensure that the multi-source tests can be compiled "by hand" without needing
additional -I options