[Kitsune] Add end-to-end-tests for Kitsune
Add end-to-end tests for Kitsune. The initial tests included here are from the
kitsune/experiments directory of the Kitsune repo. These have been divided into
three categories:

- Benchmarks
- SingleSource
- MultiSource

The benchmarks are single-source tests consisting of multiple implementations in
a range of languages. These include cuda, hip, Kokkos (with simple parallel_for
constructs), Kokkos with DualView's in addition to the Kitsune-specific forall.
These are intended to be used for both correctness checks and performance
comparisons. The implementations share as much code as possible - mostly to
read command line arguments. As a result the software engineering is somewhat
dreadful, but the goal was to enable easy experimentation by providing
convenient command line arguments to change the problem size across all of the
implementations. Each of the tests have two different problem sizes associated
with it. One is very small and executes in under a second to allow for quick
correctness checks. A larger problem is more suitable for performance
comparisons.

The tests in SingleSource/ are exactly what the name suggests - simple code
contained within a single file. They are intended to be used purely for test for
end-to-end correctness. The tests in kitsune's repository now perform many of
the same tests but in a piecemeal fashion. However, the tests in SingleSource/
are still useful since act as a sanity check over and above the granular tests.

The tests in MultiSource/ are intended to test separate compilation and LTO.
After the benchmarks, these are, arguably, the most useful since they can be
used to check "real" code that is hard to do in the core LLVM tests.

Currently, there is no support for Fortran, although some hooks have been added.
It is unlikely that we can just add Fortran files into the mix and have them
just work.

A simple timing class and another that performs floating point comparison have
been added. These could almost certainly do with improvement - especially the
timing class, but for now, they will do.

Some scripts have been provided that collect compilation and execution
statistics, both from the default report generated by the LLVM test-suite
infrastructure itself, the timings reported by the tests themselves and from
the timeit script that wraps all compilations in the test suite. This produces
a report in JSON format, but that is not the most user-friendly. Over time,
more scripts should be added that can parse that report and generate graphs or
other summaries that focus on specific characteristics.

This is also intended to supersede the code in kitsune/experiments. To that
end, some code has been imported from there that is not actually used yet. This
is intentional.

This commit also adds CPU and GPU reference outputs for the benchmarks. These
are different because the floating point error is different in the two cases.
This seems to be a function of both the frontend and backend optimizations
carried out by the compiler on these platforms. It is not clear that we can
rely on a single "expected" output for some of the more complex code.

This is a squash of commits spanning several months. There didn't seem to be
any reason to keep all those individual commits around since there was a fair
bit of churn to keep up with the significant changes taking place in parallel
in Kitsune's frontend.