[Kitsune] Fix bug that caused failures on AArch64
Assigning an integer value to a variable of type char behaves differently on
AArch64 since char's are unsigned on the latter. When assigning -1 to a char
results in the char having the value 255 on AArch64 while it is -1 on X86. The
code relied on the variable having the value -1. Changing the type of the
variable to an int, thereby matching the type of the RHS, fixes the issue.