[Kitsune] Improve flexibility of the build system
Use the list of enabled tapir targets to construct the variables that determine
which tapir targets are to be tested. This ensures that we only have to add
explicit support for a tapir target if it needs special processing. Otherwise,
it will automatically pick/discard any tapir targets that may be added to, or
removed from, Kitsune.